requirements.txt
lxml
bottlenose
python-dateutil
requests
wapy
ebaysdk
gunicorn
django-heroku
django-allauth